SD Card and car drives

+++ projects/telemetry/src/datagram.c
@@ -26,8 +26,8 @@ void log_decoded_message(Datagram *datagram) {
   LOG_DEBUG("Datagram DLC: 0x%0x\n", datagram->dlc);
 
   LOG_DEBUG("Data:\n");
-  for (size_t i = 0; i < datagram->dlc; ++i) {
+  for (uint8_t i = 0; i < datagram->dlc; ++i) {
     delay_ms(5);
-    LOG_DEBUG("Byte %d: 0x%0x\n", i, datagram->data[i]);
+    LOG_DEBUG("Byte %d: 0x%0x\n", i, (uint8_t)datagram->data[i]);
   }
 }
